backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4
requests==2.31.0
gdown==5.1.0
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2
